Spike how toolchain support could look like
Relevant bits are in JavaToolchainCompileIntegrationTest
Basic idea is that the toolchain is configured on the JavaPluginExtension

Core tasks that require the toolchain (javac, javadoc, ) expose a toolchain property and add a convention to the extension.
Externals tasks (e.g. jlink) can do the same.

Build authors are able to add more tasks and set the toolchain accordingly (API is still missing to get/add the toolchains)